+++ pkg/wharfyml/logger.go
@@ -1,5 +1,0 @@
-package wharfyml
-
-import "github.com/iver-wharf/wharf-core/pkg/logger"
-
-var log = logger.New()