+++ pkg/wharfyml/stage_test.go
@@ -4,20 +4,22 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
-func TestParseStageEnvironments(t *testing.T) {
-	content := []interface{}{"dev", "stage", "prod"}
-
-	environments, err := parseStageEnvironments(content)
-	require.Nil(t, err)
-	assert.ElementsMatch(t, content, environments)
+func TestVisitStage_ErrIfNotMap(t *testing.T) {
+	_, errs := visitStageNode(getKeyedNode(t, `myStage: 123`))
+	requireContainsErr(t, errs, ErrInvalidFieldType)
 }
 
-func TestParseStageEnvironmentsFails(t *testing.T) {
-	content := []interface{}{"dev", 1, "prod"}
+func TestVisitStage_ErrIfEmptyMap(t *testing.T) {
+	_, errs := visitStageNode(getKeyedNode(t, `myStage: {}`))
+	requireContainsErr(t, errs, ErrStageEmpty)
+}
 
-	_, err := parseStageEnvironments(content)
-	require.NotNil(t, err)
+func TestVisitStage_Name(t *testing.T) {
+	stage, errs := visitStageNode(getKeyedNode(t, `myStage: {}`))
+	if len(errs) > 0 {
+		t.Logf("errs: %v", errs)
+	}
+	assert.Equal(t, "myStage", stage.Name)
 }